+++ Saturn/src/Saturn/ImGui/TitleBar.cpp
@@ -64,7 +64,7 @@ namespace Saturn {
 			{
 				m_Height = height;
 
-				pWindow->SetTiltebarHeight( height );
+				pWindow->SetTiltebarHeight( (uint32_t)height );
 			}
 
 			pWindow->SetTitlebarCondition( ImGui::IsAnyItemHovered() );
